refactor huge debugger.js into separate files & classes
